pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tokeneyes"
version = "0.1.0"
description = "See the world in AI tokens. Point at anything, get its price in tokens."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "disc0nnctd" }]
keywords = ["ai", "tokens", "ocr", "pricing", "gemini", "claude", "openrouter", "fun"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.1",
    "google-genai>=1.0",
    "Pillow>=10.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "openai>=1.0",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov"]

[project.scripts]
tokeneyes      = "tokeneyes.cli:main"
tokeneyes-web  = "tokeneyes.web_runner:serve"

[tool.setuptools.packages.find]
where = ["."]
include = ["tokeneyes*"]

[project.urls]
Homepage = "https://github.com/disc0nnctd/TokenEyes"
Issues = "https://github.com/disc0nnctd/TokenEyes/issues"
